+++ lntest/itest/test_list_off.go
@@ -1,0 +1,5 @@
+// +build !rpctest
+
+package itest
+
+var testsCases = []*testCase{}
